[WFLY-16833]: Upgrade Apache Artemis to 2.25.0
* Upgrade to Artemis 2.25
* Return type changed from int to long.

+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/messaging/mgmt/AddressControlManagementTestCase.java
@@ -149,7 +149,7 @@ public class AddressControlManagementTestCase {
 
         Assert.assertEquals(ModelType.OBJECT, result.get("role").getType());
 
-        Assert.assertEquals(ModelType.INT, result.get("number-of-pages").getType());
+        Assert.assertEquals(ModelType.LONG, result.get("number-of-pages").getType());
 
         Assert.assertEquals(ModelType.LONG, result.get("number-of-bytes-per-page").getType());
         Assert.assertEquals(ModelType.LIST, result.get("binding-names").getType());
